+++ components/home-page/components/index.ts
@@ -1,2 +1,4 @@
 export { default as HeroAnimated } from './HeroAnimated';
-export { default as BgGradient } from './BgGradient'; 
+export { default as BgGradient } from './BgGradient';
+export { default as LinkItem } from './LinkItem';
+export { GlowyButton } from './GlowyButton';
