+++ components/home-page/icons.tsx
@@ -1,0 +1,52 @@
+import { 
+  Github,
+  Workflow,
+  ArrowRight,
+  BarChart2,
+  Activity,
+  SquareStack,
+  Waves,
+  TrendingUp,
+  Target,
+  Sparkles,
+  Truck,
+  ShieldCheck
+} from 'lucide-react';
+
+// Landing Page Icons
+export const IconWorkflow = Workflow;
+export const IconArrowRight = ArrowRight;
+export const IconBarChart2 = BarChart2;
+export const IconActivity = Activity;
+export const IconTruck = Truck;
+export const IconShieldCheck = ShieldCheck;
+
+// Timeline Steps Icons
+export const IconSquareStack = SquareStack;
+export const IconWaves = Waves;
+export const IconTrendingUp = TrendingUp;
+export const IconTarget = Target;
+export const IconSparkles = Sparkles;
+
+// Layout Icons
+export const IconGithub = Github;
+
+export default {
+  // Landing Page Icons
+  IconWorkflow,
+  IconArrowRight,
+  IconBarChart2,
+  IconActivity,
+  IconTruck,
+  IconShieldCheck,
+  
+  // Timeline Steps Icons
+  IconSquareStack,
+  IconWaves,
+  IconTrendingUp,
+  IconTarget,
+  IconSparkles,
+  
+  // Layout Icons
+  IconGithub,
+};
