+++ components/home-page/index.ts
@@ -2,10 +2,18 @@
 export { TimelineSteps } from './sections/timeline-steps';
 export { default as AnimatedFeatureCard } from './sections/animated-feature-card';
 export { default as Hero } from './sections/hero';
+export { HeroGeometric } from './sections/hero-geometric';
 
 // Home Page Layout
 export { LandingHeader } from './layout/landing-header';
 export { Footer } from './layout/footer';
 
 // Home Page Components
-export { HeroAnimated, BgGradient } from './components'; 
+export { HeroAnimated, BgGradient, LinkItem, GlowyButton } from './components';
+
+// Home Page Icons
+export * from './icons';
+
+// Home Page Utils
+export { ThemeToggle } from './layout/theme-toggle';
+export { default as SignoutButton } from './layout/signout-button';
