+++ components/home-page/layout/index.ts
@@ -1,2 +1,4 @@
 export { LandingHeader } from './landing-header';
-export { Footer } from './footer'; 
+export { Footer } from './footer';
+export { ThemeToggle } from './theme-toggle';
+export { default as SignoutButton } from './signout-button';
