+++ components/home-page/sections/index.ts
@@ -1,3 +1,4 @@
 export { TimelineSteps } from './timeline-steps';
 export { default as AnimatedFeatureCard } from './animated-feature-card';
-export { default as Hero } from './hero'; 
+export { default as Hero } from './hero';
+export { HeroGeometric } from './hero-geometric';
