+++ package.json
@@ -61,15 +61,16 @@
     "lodash": "^4.17.21",
     "lodash.debounce": "^4.0.8",
     "lucide-react": "^0.454.0",
+    "media-chrome": "^1.7.0",
     "mem0ai": "^2.1.31",
     "next": "15.2.4",
     "next-themes": "latest",
     "nuqs": "^2.4.3",
     "papaparse": "^5.5.2",
-    "react": "^19",
+    "react": "^18.2.0",
     "react-circle-flags": "^0.0.23",
     "react-day-picker": "^9.7.0",
-    "react-dom": "^19",
+    "react-dom": "^18.2.0",
     "react-draggable": "^4.4.6",
     "react-hook-form": "^7.55.0",
     "react-resizable-panels": "^2.1.7",
@@ -90,8 +91,8 @@
     "@types/lodash": "^4.17.16",
     "@types/lodash.debounce": "^4.0.9",
     "@types/node": "^22",
-    "@types/react": "^19",
-    "@types/react-dom": "^19",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "postcss": "^8",
     "tailwindcss": "^3.4.17",
     "typescript": "^5"
